Add HTTP-error filter and logging interceptor

+++ src/idea/services/idea.service.ts
@@ -1,4 +1,4 @@
-import {Injectable} from '@nestjs/common';
+import {HttpException, HttpStatus, Injectable} from '@nestjs/common';
 import {Repository} from "typeorm";
 import {IdeaEntity} from "../models/idea.entity";
 import {InjectRepository} from '@nestjs/typeorm';
@@ -24,17 +24,29 @@ export class IdeaService {
     }
 
     public async read(id: string) {
-        return await this.ideaRepository.findOne({where: {id}});
+        const idea = await this.ideaRepository.findOne({where: {id}});
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
+        return idea
     }
 
     public async update(id: string, data: Partial<IdeaDto>) {
+        const idea = await this.ideaRepository.findOne({where: id});
+        if (!idea) {
+            throw new HttpException('Cannot found idea to update', HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.update({id}, data);
-        return await this.ideaRepository.findOne({id});
+        return idea;
     }
 
     public async delete(id: string) {
+        const idea = await this.ideaRepository.findOne({where: id});
+        if (!idea) {
+            throw new HttpException('Cannot found idea to delete', HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.delete({id});
-        return {deleted: true};
+        return idea;
     }
 
 }
